Add AndroidLogcatActivityManager
