+++ com.unity.mobile.android-logcat/Editor/AndroidLogcatActivityManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 698b5bbd0008e434f9cb23105c117b1a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
